+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Version 0.4.1 - TBD
 
+### Added
+- Support for `array of` syntax
+
+
 ## Version 0.4.0 - 2023-07-06
 ### Added
 - Support for enums when they are defined separately (not inline in the property type of an entity)